requirements.txt
discord.py == 1.7.3
beautifulsoup4 == 4.9.3
requests == 2.27.1